% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concatenation.R
\name{-.tt_}
\alias{-.tt_}
\title{Concatenate textab blocks horizontally (side-by-side).}
\usage{
\method{-}{tt_}(top_row, bottom_row)
}
\arguments{
\item{top_row}{The top tabular row.}

\item{bottom_row}{The bottom tabular row.}
}
\description{
Concatenate textab blocks horizontally (side-by-side).
}
