% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concatenation.R
\name{print.tt_block}
\alias{print.tt_block}
\title{Print a textab block as a LaTeX tabular.}
\usage{
\method{print}{tt_block}(x, ...)
}
\arguments{
\item{x}{A textab block.}

\item{...}{Placeholder for print.}
}
\value{
Print prints its argument and returns it invisibly.
}
\description{
Print a textab block as a LaTeX tabular.
}
