% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/midrule.R
\name{TexMidrule}
\alias{TexMidrule}
\title{Create one (or many) partial midrule(s).}
\usage{
TexMidrule(rule_list = NULL)
}
\arguments{
\item{rule_list}{(list). A list of integer vectors. Each integer vector must contain two integers which indicate the start and end column of the partial midrule. If rule_list = NULL, it returns the full midrule across all columns. The default is rule_list = NULL.}
}
\description{
Create one (or many) partial midrule(s).
}
