% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concatenation.R
\name{+.tt_}
\alias{+.tt_}
\title{Concatenate textab blocks vertically.}
\usage{
\method{+}{tt_}(left_block, right_block)
}
\arguments{
\item{left_block}{The left block of the tabular row.}

\item{right_block}{The right block of the tabular row.}
}
\value{
A block of a tabular, formed by vertically concatenating the two provided blocks.
}
\description{
Concatenate textab blocks vertically.
}
